feat: DTO 데이터타입 수정

+++ src/main/java/nbcamp/food_order_platform/product/presentation/dto/PostProductReqDto.java
@@ -8,8 +8,8 @@ import java.util.UUID;
 public class PostProductReqDto {
     private UUID storeId;
     private String name;
-    private Long price;
-    private Integer stockQuantity;
+    private int price;
+    private int stockQuantity;
     private boolean useAi;
     private String description;
 }
